# openapi/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/files/presign-upload": {
      "post": {
        "operationId": "FilesController_presign",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/files/confirm": {
      "post": {
        "operationId": "FilesController_confirm",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/files/{attachmentId}/download-url": {
      "get": {
        "operationId": "FilesController_download",
        "parameters": [
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/files/{attachmentId}": {
      "delete": {
        "operationId": "FilesController_remove",
        "parameters": [
          {
            "name": "attachmentId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "files"
        ],
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "SSERP API",
    "description": "Special School & Therapy Center ERP",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "PresignDto": {
        "type": "object",
        "properties": {}
      },
      "ConfirmDto": {
        "type": "object",
        "properties": {}
      }
    }
  }
}